pyproject.toml
[tool.poetry]
name = "fastapi-simple-rate-limiter"
version = "0.0.1"
description = "Rate limiter to limit the number of API requests in FastAPI"
authors = ["jintaekimmm <jintae.kimmm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
redis = "^5.0.1"

[tool.poetry.group.test.dependencies]
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

